backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg[binary]==3.2.12
celery==5.6.0
redis==6.4.0
elasticsearch==9.3.0
Pillow==11.2.1
gunicorn==23.0.0
